backend/server/requirements.txt
fastapi==0.70.0
pandas==1.3.3
py2neo==2021.2.3
PyYAML==6.0
uvicorn==0.14.0
